Factor out hotdocs_xml_parsing module.
