+++ hpaction/docusign.py
@@ -1,7 +1,5 @@
-from typing import List, Optional, NamedTuple, Union
-from enum import Enum
+from typing import List, Optional, NamedTuple
 from io import BytesIO
-import xml.etree.ElementTree as ET
 import base64
 import logging
 from django.http import (
@@ -19,6 +17,7 @@ from docusign.core import docusign_client_user_id
 from docusign.views import create_callback_url, append_querystring_args
 from onboarding.models import BOROUGH_CHOICES
 import docusign_esign as dse
+from .hotdocs_xml_parsing import HPAType
 from .models import HPActionDocuments, DocusignEnvelope, HP_DOCUSIGN_STATUS_CHOICES, Config
 
 
@@ -37,37 +36,6 @@ NUM_COVER_SHEET_PAGES = 1
 
 
 logger = logging.getLogger(__name__)
-
-
-def get_answers_xml_tf(root: ET.Element, name: str) -> Optional[bool]:
-    nodes = root.findall(f".//Answer[@name='{name}']/TFValue")
-    if nodes:
-        return nodes[0].text == 'true'
-    return None
-
-
-class HPAType(Enum):
-    REPAIRS = 1
-    HARASSMENT = 2
-    BOTH = 3
-
-    @staticmethod
-    def get_from_answers_xml(xml_value: Union[str, bytes]) -> 'HPAType':
-        # Interestingly, ET is in charge of decoding this if it's bytes:
-        # https://stackoverflow.com/a/21698118
-        root = ET.fromstring(xml_value)
-
-        harassment = get_answers_xml_tf(root, 'Sue for harassment TF')
-        repairs = get_answers_xml_tf(root, 'Sue for repairs TF')
-
-        if harassment and repairs:
-            return HPAType.BOTH
-        elif harassment:
-            return HPAType.HARASSMENT
-        elif repairs:
-            return HPAType.REPAIRS
-
-        raise ValueError('XML is suing for neither harassment nor repairs!')
 
 
 class PageCoords(NamedTuple):
