+++ hpaction/tests/test_docusign.py
@@ -16,7 +16,6 @@ from loc.tests.factories import LandlordDetailsFactory
 from hpaction.models import Config
 from hpaction import docusign
 from hpaction.docusign import HPAType, FormsConfig
-from hpaction.hpactionvars import HPActionVariables
 
 ALL_BOROUGHS = BOROUGH_CHOICES.choices_dict.keys()
 
@@ -26,35 +25,6 @@ def set_config(**kwargs):
     for name, value in kwargs.items():
         setattr(config, name, value)
     config.save()
-
-
-class TestHPAType:
-    @pytest.mark.parametrize('vars,expected', [
-        (HPActionVariables(sue_for_harassment_tf=True), HPAType.HARASSMENT),
-        (HPActionVariables(sue_for_harassment_tf=True,
-                           sue_for_repairs_tf=False), HPAType.HARASSMENT),
-        (HPActionVariables(sue_for_repairs_tf=True), HPAType.REPAIRS),
-        (HPActionVariables(sue_for_harassment_tf=False,
-                           sue_for_repairs_tf=True), HPAType.REPAIRS),
-        (HPActionVariables(sue_for_repairs_tf=True,
-                           sue_for_harassment_tf=True), HPAType.BOTH),
-    ])
-    def test_it_works(self, vars: HPActionVariables, expected):
-        xmlstr = str(vars.to_answer_set())
-        assert HPAType.get_from_answers_xml(xmlstr) == expected
-
-        vars.access_person_te = "with unicode\u2026"
-        xmlbytes = str(vars.to_answer_set()).encode('utf-8')
-        assert HPAType.get_from_answers_xml(xmlbytes) == expected
-
-    @pytest.mark.parametrize('vars', [
-        HPActionVariables(),
-        HPActionVariables(sue_for_harassment_tf=False, sue_for_repairs_tf=False),
-    ])
-    def test_it_raises_error_when_neither_are_present(self, vars):
-        xmlstr = str(vars.to_answer_set())
-        with pytest.raises(ValueError, match="suing for neither"):
-            HPAType.get_from_answers_xml(xmlstr)
 
 
 class TestGetHousingCourtForBorough:
